[course 4] examples
